Tried to comply to the XDG specifications.
The trash directory by default is
$XDG_DATA_HOME/remTrash. If XDG_DATA_HOME is not
set, the choice of the trash directory is made in
a smart-ish way.
I know that I will definitively not use this
software if it is going to add clutter to my home
directory and I might not be the only one in that
case. Following the XDG specification could let
REM be used by more people.